+++ CHANGELOG.md
@@ -7,15 +7,18 @@
 ### New features
 
 * Add `kala.control.OptionContainer::getOrThrow()`
-* Add `kala.collection.base.OrderedTraversable::mapIndexed(CollectionFactory<U, ?, R>, IndexedFunction<T, U>)`
-* Add `kala.collection.base.OrderedTraversable::mapIndexedTo(G, IndexedFunction<T, U>)`
 * Add `kala.collection.MapLike::getOrThrow(K)`
+* Add `kala.collection.SeqLike::getOrDefault(int, E)`
+* Add `kala.collection.SeqLike::getOrElse(int, Supplier<E>)`
+* Add `kala.collection.SeqLike::getOrThrow(int, Supplier<Ex>)`
 * Add `kala.collection.Set::added(E)`
 * Add `kala.collection.Set::addedAll(E...)`
 * Add `kala.collection.Set::addedAll(Iterable<E>)`
 * Add `kala.collection.SetLike::removed(E)`
 * Add `kala.collection.SetLike::removedAll(E...)`
 * Add `kala.collection.SetLike::removedAll(Iterable<E>)`
+* Add `kala.collection.base.OrderedTraversable::mapIndexed(CollectionFactory<U, ?, R>, IndexedFunction<T, U>)`
+* Add `kala.collection.base.OrderedTraversable::mapIndexedTo(G, IndexedFunction<T, U>)`
 * Add `kala.function.Predicates.isNotEquals(T)`
 * Add `kala.function.Predicates.isNotSame(T)`